feat(rs): solution for 2021 day 24
This is a bit disappointing in that it does not really provide a neat
algorithmic solution to find the desired answers:  I have done a manual
search for the correct numbers after spending way too long on figuring
out the right angle of attack.  But at least the (entirely pointless)
ALU implementation was fun.