chore(rs): fix-up test for 2022 day 19 part 2
Test didn't actually work (my input happened to give the right answer)
and on top of that was way too slow and blowing memory on the cache.
Adding some clamping in the resource tracking to get better memoisation
usage, but the clamping is hand-tuned to my input and the test, not
derived from principle.