feat(rs): improve runtime of 2023 day 6
Instead of iteration, just calculate the allowed range directly via the
analytic solution.  Slight complication with int/float conversions, so
ensuring that the first guess for the range is larger if not exact and
then adjust boundary points.

This is entirely unnecessary for the problem, but it does run 100x
faster (from 30ish ms to 300ish μs).